# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RHC API Documentation",
        "description": "API documentation for the RHC backend built with Laravel",
        "contact": {
            "email": "support@rhc.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.rhc.com/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/book-appointment": {
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Book a new appointment",
                "operationId": "deee02e5516de93f2892a980c8d49387",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone",
                                    "unit_service",
                                    "datetime",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+1234567890"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "dob": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "unit_service": {
                                        "type": "string",
                                        "example": "General Checkup"
                                    },
                                    "datetime": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-12-01 10:00:00"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "in_person",
                                            "online",
                                            "follow_up"
                                        ],
                                        "example": "in_person"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Additional notes about the appointment"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Appointment booked successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Appointment booked successfully"
                                        },
                                        "confirmation_code": {
                                            "type": "string",
                                            "example": "RHC-20231201-ABC123"
                                        },
                                        "appointment_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "The name field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/appointments/{id}": {
            "put": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Update an appointment (admin only)",
                "operationId": "ac9d265009479404ff79f23602c22aca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Appointment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "confirmed",
                                            "cancelled",
                                            "completed"
                                        ]
                                    },
                                    "datetime": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-12-01 10:00:00"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Updated notes"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Appointment updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Appointment updated successfully"
                                        },
                                        "appointment": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "patient_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "unit_service": {
                                                    "type": "string",
                                                    "example": "General Checkup"
                                                },
                                                "datetime": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "in_person"
                                                },
                                                "notes": {
                                                    "type": "string",
                                                    "example": "Updated notes"
                                                },
                                                "confirmation_code": {
                                                    "type": "string",
                                                    "example": "RHC-20231201-ABC123"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "confirmed"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "patient": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "John Doe"
                                                        },
                                                        "phone": {
                                                            "type": "string",
                                                            "example": "+1234567890"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "format": "email",
                                                            "example": "john@example.com"
                                                        },
                                                        "dob": {
                                                            "type": "string",
                                                            "format": "date",
                                                            "example": "1990-01-01"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Appointment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Appointment not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "status": [
                                                    "The selected status is invalid."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}